Entrega final de todo

+++ tallersena .NET/tallersena/Modelo/Estudiantes.cs
@@ -13,7 +13,7 @@ namespace tallersena.Modelo
         public double? Nota2 { get; set; }
         public double? Nota3 { get; set; }
 
-        public double NotaFinal(int N1, int N2, int N3)
+        public double NotaFinal(double N1, double N2, double N3) 
         {
             double NF = (N1 + N2 + N3) / 3;
             return NF;
